Allow customization of installation directories
Add a LIBDIR variable, defaulting to $(PREFIX)/lib, to allow overriding
to /usr/lib64 on aarch64.

Also add a MANDIR variable, defaulting to $(PREFIX)/man, to allow
overriding to /usr/share/man.

For completeness add BINDIR and INCLUDEDIR variables as well.

In an RPM environment the expected use case is:
 BINDIR=%_bindir INCLUDEDIR=%_includedir LIBDIR=%_libdir MANDIR=%_mandir

This fixes the installation locations for openSUSE.

Signed-off-by: Andreas Färber <afaerber@suse.de>